[buck2]UPDATE: Encapsulate tasks and targets model in orion-server
