+++ jupiter/callisto/src/build_targets.rs
@@ -1,6 +1,6 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
 
-use sea_orm::{IntoActiveModel, entity::prelude::*};
+use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
@@ -41,29 +41,6 @@ impl Related<super::target_state_histories::Entity> for Entity {
 }
 
 impl ActiveModelBehavior for ActiveModel {}
-
-impl Model {
-    pub fn create_default_target(id: Uuid, task_id: Uuid) -> Self {
-        let default_path = "//";
-        Self {
-            id,
-            task_id,
-            path: serde_json::to_value(default_path).unwrap(),
-            target_state: "default".to_string(),
-        }
-    }
-
-    pub async fn insert_default_target(
-        id: Uuid,
-        task_id: Uuid,
-        db: &impl ConnectionTrait,
-    ) -> Result<String, DbErr> {
-        let target = Self::create_default_target(id, task_id);
-        let path = target.path.clone();
-        target.into_active_model().insert(db).await?;
-        Ok(path.to_string())
-    }
-}
 //:path1 //:path2 //:path3
 
 //target log:
