+++ jupiter/callisto/src/orion_tasks.rs
@@ -1,9 +1,7 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
 
-use api_model::buck2::{status::Status, types::ProjectRelativePath};
-use sea_orm::{IntoActiveModel, entity::prelude::*};
+use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
-use serde_json::to_value;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
 #[sea_orm(table_name = "orion_tasks")]
@@ -46,32 +44,3 @@ impl Related<super::target_build_status::Entity> for Entity {
 }
 
 impl ActiveModelBehavior for ActiveModel {}
-
-impl Model {
-    fn create_task(
-        task_id: Uuid,
-        cl_link: &str,
-        repo: &str,
-        changes: &Vec<Status<ProjectRelativePath>>,
-    ) -> Result<Self, serde_json::Error> {
-        Ok(Self {
-            id: task_id,
-            cl: cl_link.to_string(),
-            repo_name: repo.to_string(),
-            changes: to_value(changes)?,
-            created_at: chrono::Utc::now().into(),
-        })
-    }
-
-    pub async fn insert_task(
-        task_id: Uuid,
-        cl_link: &str,
-        repo: &str,
-        changes: &Vec<Status<ProjectRelativePath>>,
-        db: &impl ConnectionTrait,
-    ) -> Result<Model, DbErr> {
-        let task_model = Self::create_task(task_id, cl_link, repo, changes)
-            .map_err(|e| DbErr::Custom(e.to_string()))?;
-        task_model.into_active_model().insert(db).await
-    }
-}
