+++ orion-server/src/model/build_targets.rs
@@ -1,43 +1,27 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
+use sea_orm::{ActiveModelTrait, ConnectionTrait, IntoActiveModel};
 
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
+/// A collection of utility methods for the `build_targets` database table.
+pub struct BuildTarget;
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(table_name = "build_targets")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    pub task_id: Uuid,
-    pub path: String,
-    #[sea_orm(column_type = "Text")]
-    pub latest_state: String,
-}
+impl BuildTarget {
+    pub fn create_default_target(id: Uuid, task_id: Uuid) -> callisto::build_targets::Model {
+        let default_path = "//";
+        callisto::build_targets::Model {
+            id,
+            task_id,
+            path: default_path.to_string(),
+            latest_state: "NOT_STARTED".to_string(),
+        }
+    }
 
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::orion_tasks::Entity",
-        from = "Column::TaskId",
-        to = "super::orion_tasks::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    OrionTasks,
-    #[sea_orm(has_many = "super::target_state_histories::Entity")]
-    TargetStateHistories,
-}
-
-impl Related<super::orion_tasks::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::OrionTasks.def()
+    pub async fn insert_default_target(
+        id: Uuid,
+        task_id: Uuid,
+        db: &impl ConnectionTrait,
+    ) -> Result<String, DbErr> {
+        let target = Self::create_default_target(id, task_id);
+        let path = target.path.clone();
+        target.into_active_model().insert(db).await?;
+        Ok(path.to_string())
     }
 }
-
-impl Related<super::target_state_histories::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::TargetStateHistories.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
